+++ src/main.ts
@@ -4,12 +4,23 @@
 import { Plugin } from "obsidian";
 import { DEFAULT_SETTINGS, normalizeSettings, ScribeGoalsStatsSettings } from "./settings/settings";
 import { ScribeGoalsStatsSettingTab } from "./settings/settingsTab";
+import { ExplorerDecorator } from "./views/explorerDecorator";
 
 export default class ScribeGoalsStatsPlugin extends Plugin {
 	settings: ScribeGoalsStatsSettings = { ...DEFAULT_SETTINGS };
+	private explorerDecorator!: ExplorerDecorator;
 
 	async onload(): Promise<void> {
 		await this.loadSettings();
+
+		this.explorerDecorator = this.addChild(
+			new ExplorerDecorator(this.app, () => ({
+				storyFolder: this.settings.storyFolder,
+				excludedPaths: this.settings.excludedPaths,
+				metric: this.settings.metric,
+			})),
+		);
+
 		this.addSettingTab(new ScribeGoalsStatsSettingTab(this.app, this));
 	}
 
@@ -21,5 +32,6 @@ export default class ScribeGoalsStatsPlugin extends Plugin {
 
 	async saveSettings(): Promise<void> {
 		await this.saveData(this.settings);
+		void this.explorerDecorator.refresh();
 	}
 }